Revert "Merge branch 'master' of https://github.com/rvielmaa/Auto-GPT"
This reverts commit 81da0c3a844f12b4902481a430d71d7a9a8872aa, reversing
changes made to 236b17d2f2299207b69911e818354a6f82d01ced.
